+++ python/tests/sync_tests/conftest.py
@@ -137,6 +137,7 @@ def create_sync_client(
     lazy_connect: Optional[bool] = False,
     enable_compression: Optional[bool] = None,
     inflight_requests_limit: Optional[int] = None,
+    read_only: bool = False,
 ) -> TSyncGlideClient:
     # Create sync client
     config = create_sync_client_config(
@@ -160,6 +161,7 @@ def create_sync_client(
         lazy_connect=lazy_connect,
         enable_compression=enable_compression,
         inflight_requests_limit=inflight_requests_limit,
+        read_only=read_only,
     )
     if cluster_mode:
         return SyncGlideClusterClient.create(config)